packages/sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rifft"
version = "0.1.0"
description = "Cross-framework debugger for multi-agent AI systems"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
  "opentelemetry-sdk>=1.30.0",
  "opentelemetry-exporter-otlp>=1.30.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

